requirements.txt
# ============================================================
# GuildPilot - Consolidated requirements.txt
# One bot process + one venv. Dependencies grouped by module.
# Discord library choice: py-cord (ONLY). Do not install discord.py too.
# ============================================================


# ============================
# Core / Shared Dependencies
# ============================
py-cord==2.6.1
python-dotenv==1.0.1
aiohttp==3.11.11
requests==2.32.3

# Shared networking / aiohttp stack
aiohappyeyeballs==2.4.4
aiosignal==1.3.2
attrs==24.3.0
frozenlist==1.5.0
multidict==6.1.0
propcache==0.2.1
yarl==1.18.3

# Shared HTTP / parsing
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
typing_extensions==4.12.2


# ============================
# Module: PilotAI
# (AI chatbot + API client)
# ============================
openai==0.28.0
httpx==0.27.2
httpcore==1.0.5
h11==0.14.0
anyio==4.4.0
sniffio==1.3.1
tqdm==4.66.5
distro==1.9.0
jiter==0.5.0
annotated-types==0.7.0
pydantic==2.8.2
pydantic_core==2.20.1




# ============================
# Module: RoleCop
# (member assignment / permissions / voice, etc.)
# ============================
PyNaCl==1.5.0
cffi==1.17.0
pycparser==2.22
pytz==2024.1
DateTime==5.5
zope.interface==7.0.3


# ============================
# Module: StatWrangler
# (game stats lookup / scraping / automation)
# ============================
numpy==2.2.1
selenium==4.27.1
webdriver_manager==4.0.2
PySocks==1.7.1
validators==0.35.0

# Selenium runtime deps
trio==0.28.0
trio-websocket==0.11.1
sortedcontainers==2.4.0
outcome==1.3.0.post0
websocket-client==1.8.0
wsproto==1.2.0

# Video tooling 
ffmpeg-python==0.2.0
imageio-ffmpeg==0.5.1
future==1.0.0

# Utility
packaging==24.2


# ============================
# Module: StreamSentinel
# (Twitch → Discord announcer)
# ============================
# (uses shared: py-cord, python-dotenv, requests)
